libs/common: Refactor libs/common 01
