+++ libs/kv/result.go
@@ -1,6 +1,6 @@
 // nolint: dupl
 // dupl is reading this as the same file as crypto/merkle/result.go
-package common
+package kv
 
 import (
 	"bytes"
