+++ libs/kv/types.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
-package tendermint.libs.common;
-option go_package = "github.com/tendermint/tendermint/libs/common";
+package tendermint.libs.kv;
+option go_package = "github.com/tendermint/tendermint/libs/kv";
 
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
